机构添加，客服添加

+++ smile_admin/src/main/java/org/flightythought/smile/admin/database/repository/OfficeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OfficeRepository extends JpaRepository<OfficeEntity,Integer> {
+public interface OfficeRepository extends JpaRepository<OfficeEntity,Long> {
 }
